move process_labels_3d out
