+++ crates/re_space_view_spatial/src/visualizers/mod.rs
@@ -23,7 +23,7 @@ pub use images::ImageVisualizer;
 pub use segmentation_images::SegmentationImageVisualizer;
 pub use transform3d_arrows::{add_axis_arrows, AxisLengthDetector, Transform3DArrowsVisualizer};
 pub use utilities::{
-    bounding_box_for_textured_rect, entity_iterator, tensor_to_textured_rect,
+    bounding_box_for_textured_rect, entity_iterator, process_labels_3d, tensor_to_textured_rect,
     SpatialViewVisualizerData, UiLabel, UiLabelTarget, MAX_NUM_LABELS_PER_ENTITY,
 };
