+++ crates/re_space_view_spatial/src/visualizers/points3d.rs
@@ -1,7 +1,5 @@
 use itertools::Itertools as _;
 
-use re_entity_db::{EntityPath, InstancePathHash};
-use re_log_types::Instance;
 use re_query::range_zip_1x5;
 use re_renderer::{LineDrawableBuilder, PickingLayerInstanceId, PointCloudBuilder};
 use re_types::{
@@ -10,8 +8,8 @@ use re_types::{
 };
 use re_viewer_context::{
     auto_color_for_entity_path, ApplicableEntities, IdentifiedViewSystem, QueryContext,
-    ResolvedAnnotationInfos, SpaceViewSystemExecutionError, TypedComponentFallbackProvider,
-    ViewContext, ViewContextCollection, ViewQuery, VisualizableEntities, VisualizableFilterContext,
+    SpaceViewSystemExecutionError, TypedComponentFallbackProvider, ViewContext,
+    ViewContextCollection, ViewQuery, VisualizableEntities, VisualizableFilterContext,
     VisualizerQueryInfo, VisualizerSystem,
 };
 
@@ -20,12 +18,12 @@ use crate::{
     view_kind::SpatialSpaceViewKind,
     visualizers::{
         load_keypoint_connections, process_annotation_and_keypoint_slices, process_color_slice,
-        process_radius_slice, UiLabel, UiLabelTarget,
+        process_radius_slice,
     },
 };
 
 use super::{
-    entity_iterator::clamped, filter_visualizable_3d_entities, SpatialViewVisualizerData,
+    filter_visualizable_3d_entities, process_labels_3d, SpatialViewVisualizerData,
     MAX_NUM_LABELS_PER_ENTITY, SIZE_BOOST_IN_POINTS_FOR_POINT_OUTLINES,
 };
 
@@ -58,34 +56,6 @@ struct Points3DComponentData<'a> {
 // NOTE: Do not put profile scopes in these methods. They are called for all entities and all
 // timestamps within a time range -- it's _a lot_.
 impl Points3DVisualizer {
-    fn process_labels<'a>(
-        entity_path: &'a EntityPath,
-        positions: &'a [glam::Vec3],
-        labels: &'a [Text],
-        colors: &'a [egui::Color32],
-        annotation_infos: &'a ResolvedAnnotationInfos,
-        world_from_obj: glam::Affine3A,
-    ) -> impl Iterator<Item = UiLabel> + 'a {
-        let labels = clamped(labels, positions.len());
-        itertools::izip!(annotation_infos.iter(), positions, labels, colors)
-            .enumerate()
-            .filter_map(move |(i, (annotation_info, point, label, color))| {
-                let label = annotation_info.label(Some(label.as_str()));
-                match (point, label) {
-                    (point, Some(label)) => Some(UiLabel {
-                        text: label,
-                        color: *color,
-                        target: UiLabelTarget::Position3D(world_from_obj.transform_point3(*point)),
-                        labeled_instance: InstancePathHash::instance(
-                            entity_path,
-                            Instance::from(i as u64),
-                        ),
-                    }),
-                    _ => None,
-                }
-            })
-    }
-
     fn process_data<'a>(
         &mut self,
         ctx: &QueryContext<'_>,
@@ -173,9 +143,9 @@ impl Points3DVisualizer {
                     positions
                 };
 
-                self.data.ui_labels.extend(Self::process_labels(
+                self.data.ui_labels.extend(process_labels_3d(
                     entity_path,
-                    label_positions,
+                    label_positions.iter(),
                     data.labels,
                     &colors,
                     &annotation_infos,
