+++ crates/re_space_view_spatial/src/visualizers/utilities/mod.rs
@@ -3,6 +3,6 @@ mod labels;
 mod spatial_view_visualizer;
 mod textured_rect;
 
-pub use labels::{UiLabel, UiLabelTarget, MAX_NUM_LABELS_PER_ENTITY};
+pub use labels::{process_labels_3d, UiLabel, UiLabelTarget, MAX_NUM_LABELS_PER_ENTITY};
 pub use spatial_view_visualizer::SpatialViewVisualizerData;
 pub use textured_rect::{bounding_box_for_textured_rect, tensor_to_textured_rect};
